# fix performance regression when using `field_split` and `value_split` char classes