@colinsurprenant 2018-09-04
Good strategy! LGTM.